Add magnification factor for exporting quilts
Exporting quilt images/movies with the default display window size
seems to produce lower resolution quilt image files. When viewing these
quilt image files in an LG device, they do not appear as high of
resolution as what was displayed in ParaView.

Adding a magnification factor for the display window fixes this issue,
though, by increasing the size of the images that get written.

A magnification factor of 2 appears to give pretty good results. But
the magnification factor can also now be set in the interface in case
the user wants to use a different magnification factor.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>