Put common render code in a RenderQuilt() function
This moves common rendering code into a
`vtkLookingGlassInterface::RenderQuilt()` function. The common rendering
code is being moved from three places:

1. `vtkLookingGlassPass::Render()`
2. `vtkLookingGlassRenderWindowImpl::DoStereoRender()`
3. The ParaView plugin

An optional `renderFunc` argument was added specifically for the
`vtkLookingGlassPass` case, where instead of calling `Render()` on the
renderers as the other two cases do, it needs to call `Render()` on its
`DelegatePass` with a separate `vtkRenderState*` argument.

I tried to also look into resizing the render window (to match the
`RenderSize`) inside the `RenderQuilt()` function. This would work for
the ParaView plugin, which has separate render windows for rendering
the quilt and rendering to the display, but:

1. For the `vtkLookingGlassPass`, it appears that we do not want to resize
the render window. I'm not entirely sure why.
2. For the `vtkLookingGlassRenderWindowImpl`, we have to bypass `SetSize()`
on the render window, at least for Linux (which uses `vtkXOpenGLRenderWindow`),
because this otherwise causes a re-render when we do not want it, and a
resulting render flash on the display.

So for the ParaView plugin and for the render window implementation, we fix the
size ourselves outside of the `RenderQuilt()` function.

This is working in my tests for all three cases.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>