Save quilts via textures, not the render window
This fixes several issues we were encountering when saving quilts (both
images and movies) in pure VTK. Previously, quilts were saved using the
render window. One problem, though, is that this required resizing the
render window to be the quilt size, then temporarily rendering a quilt
on it, writing it out, resizing the render window to its original size,
and then re-rendering what was previously on it. The resizing of the
render window would mark it as modified, which would cause render window
interactors to resize it yet again.

Using the textures to save the quilts is a much more straightforward and
simple path that does not involve manipulation of the render window. It
also does not require any extra rendering, so it is more efficient as
well.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>