Fix import paths in python examples
When we install this module separately from the regular VTK wheel, we cannot
access classes like `vtkLookingGlassInterface` directly on the VTK module.
Instead, we have to import `vtkRenderingLookingGlass` first, and then we can
access the classes underneath that.

Fix these import paths so that they work both for a local VTK build, and for
the wheels.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>