Add wheel repair step for Linux
This does the following:

1. Copies all VTK modules into wheel
2. Runs audit wheel repair
3. Removes all files/directories that were added to the wheel

This is done so that auditwheel can add tags to the wheel without
modifying RPATHs or the wheel contents.

We have to copy the VTK modules into the wheel to avoid modifying the
RPATHs. If we simply make them visible to auditwheel, then auditwheel
copies them to a different directory and modifies the RPATHs of our
files.

If auditwheel at some point allows us to add tags without repairing the
wheel, that would be ideal. But for now, you must repair the wheel for
tags to be added correctly.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>