Add several functions to the render window
This exposes several `vtkLookingGlassInterface` functions on the render
window classes, which make them easier to access, including in Python.
These are functions that we intend for users to have easy access to.

This also adds a `GetDeviceTypes()` function that can be called from
python (apparently, `vtkLookingGlassInterface::GetDevices()` does not
get a python interface, perhaps because of its return type).

Additionally, changing devices in `vtkLookingGlassInterface` is not
straightforward, and it's easier to just make a new one from scratch.
So, when the device type switches, re-create the interface.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>