Add Linux environment variable for dev building
If this environment variable is set, then you can build a wheel locally
simply via `pip wheel . --no-deps`, and then install and use that wheel.

This enables much faster development of the C++ code in this repo, since
you don't need to rebuild anything other than this module alone for
development.

For Windows and Mac, this environment variable should not be necessary,
and it should already work. It's only needed on Linux so that we use an
older C++ ABI.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>