Add example of custom interaction style
This script provides an example of a custom interaction style that
utilizes some of the looking glass functions, including moving the
focal plane, turning on/off clipping, and moving the clipping planes.

It is most helpful for being able to interactively determine where the
focal plane and clipping planes should be located to produce the desired
image.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>