Keep track of device aspect ratios
Previously, the aspect ratio of the currently connected device would
be used when generating a quilt/image for a different device. This was
incorrect, and would make the quilts appear blurry.

Instead, keep track of the aspect ratios for the various devices, and
use that aspect ratio when generating a quilt for that device.

This, for instance, makes quilts generated for a "Portrait" device look
much nicer even though a "Large" device was connected at the time of
the quilt generation.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>