Turn off arm64 builds in cibuildwheel
Because GitHub Actions does not yet have arm64 runners available,
cross-compiling is the only option for making arm64 wheels.

We are running into several issues with cross-compiling our module,
including the fact that, apparently, part of the compilation process
runs at least one executable (vtkWrapHierarchy-9.2) in the VTK SDK
(which, if we are cross-compiling, will not work since the architecture
does not match). We also seem to be getting some linking errors, which
may or may not be important.

Instead of trying to fix the cross-compilation, let's turn it off and
wait until GitHub Actions has arm64 runners available. In the meantime,
we can build the wheels on a local arm64 machine manually and push them
to PyPI, which is not very hard.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>