Downgrade the LookingGlass SDK that Linux uses
The latest Looking Glass SDK was built with too new of a glibc version
for Linux. Downgrade it so that we can package it into wheels.

We have unfortunately been unable to determine whether the older
Looking Glass SDK will work with the new Looking Glass Bridge software.
Hopefully, it does. If not, then we will have to require Linux users to
use an older HoloPlay Service.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>